requirements.txt
anyio==4.9.0
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
clients==1.5
crispy-bootstrap4==2024.10
Django==5.0.1
django-admin-thumbnails==0.2.8
django-adminlte-3==0.1.6
django-ckeditor==6.7.1
django-crispy-forms==2.3
django-embed-video==1.4.10
django-jet-reboot==1.3.3
django-js-asset==2.2.0
django-mptt==0.16.0
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.0
razorpay==1.4.2
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.8.2
